[ctxprof][nfc] Move profile annotator to Analysis
